mastering

+++ tao_of_rust/rmX.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 for file in `ls`; do
-if [ -x "$file" ]; then
+if [ -x "$file" ] && [ -f "$file" ]; then
    if [ "$file" != "rmX.sh" ]; then
       rm -rf $file
    fi
